analytics/requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
imbalanced-learn==0.11.0

# Database
mysql-connector-python==8.2.0
sqlalchemy==2.0.23

# API
python-dotenv==1.0.0
gunicorn==21.2.0

# Utilities
joblib==1.3.2
